backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
beanie==2.0.0
    # via -r requirements.in
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   beanie
    #   rich-toolkit
    #   typer
    #   uvicorn
debugpy==1.8.16
    # via -r requirements.in
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via
    #   fastapi
    #   pydantic
fastapi[standard]==0.116.1
    # via -r requirements.in
fastapi-cli[standard]==0.0.8
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
google-auth==2.40.3
    # via google-genai
google-genai==1.33.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via fastapi
lazy-model==0.3.0
    # via beanie
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
motor==3.7.1
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic[email]==2.11.7
    # via
    #   beanie
    #   fastapi
    #   fastapi-cloud-cli
    #   google-genai
    #   lazy-model
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pygments==2.19.2
    # via rich
pymongo==4.14.1
    # via
    #   beanie
    #   motor
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
requests==2.32.5
    # via google-genai
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.0
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
rsa==4.9.1
    # via google-auth
ruff==0.12.11
    # via -r requirements.in
sentry-sdk==2.35.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via fastapi
tenacity==9.1.2
    # via google-genai
typer==0.16.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.14.1
    # via
    #   beanie
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==14.2
    # via
    #   google-genai
    #   uvicorn
